cruft update

+++ webapp/src/routes/my/logout/+page.ts
@@ -1,9 +1,12 @@
+import { browser } from '$app/environment';
 import { currentUser, pb } from '$lib/pocketbase';
 import { redirect } from '@sveltejs/kit';
 
 export const load = async () => {
-	localStorage.clear();
-	pb.authStore.clear();
-	currentUser.set(null);
-	throw redirect(303, '/');
+	if (browser) {
+		localStorage.clear();
+		pb.authStore.clear();
+		currentUser.set(null);
+		throw redirect(303, '/');
+	}
 };
